+++ permission/src/iosMain/kotlin/tech/kotlinlang/permission/IosLocationManager.kt
@@ -15,13 +15,14 @@ import platform.CoreLocation.kCLLocationAccuracyBest
 import platform.Foundation.NSError
 import platform.darwin.NSObject
 import tech.kotlinlang.permission.location.LocationRequestResult
+import tech.kotlinlang.permission.result.LocationPermissionResult
 import kotlin.coroutines.resume
 
 class IosLocationManager : NSObject(), CLLocationManagerDelegateProtocol {
 
     private val locationManager by lazy { CLLocationManager() }
 
-    private var locationPermissionContinuation: (CancellableContinuation<LocationGrantedType>)? = null
+    private var locationPermissionContinuation: (CancellableContinuation<LocationPermissionResult>)? = null
     private var locationContinuation: (CancellableContinuation<LocationRequestResult>)? = null
 
     init {
@@ -29,7 +30,7 @@ class IosLocationManager : NSObject(), CLLocationManagerDelegateProtocol {
         locationManager.desiredAccuracy = kCLLocationAccuracyBest
     }
 
-    suspend fun askForLocationPermission(): LocationGrantedType {
+    suspend fun askForLocationPermission(): LocationPermissionResult {
         return suspendCancellableCoroutine { continuation ->
             locationPermissionContinuation = continuation
             locationManager.requestWhenInUseAuthorization()
@@ -66,21 +67,21 @@ class IosLocationManager : NSObject(), CLLocationManagerDelegateProtocol {
         locationPermissionContinuation = null
     }
 
-    fun checkLocationPermission(status: CLAuthorizationStatus): LocationGrantedType {
+    fun checkLocationPermission(status: CLAuthorizationStatus): LocationPermissionResult {
         return when (status) {
             kCLAuthorizationStatusAuthorizedAlways, kCLAuthorizationStatusAuthorizedWhenInUse -> {
-                if (locationManager.accuracyAuthorization.value == 1L) LocationGrantedType.GrantedType.Approximate
-                else LocationGrantedType.GrantedType.Precise
+                if (locationManager.accuracyAuthorization.value == 1L) LocationPermissionResult.Granted.Approximate
+                else LocationPermissionResult.Granted.Precise
             }
 
-            kCLAuthorizationStatusRestricted, kCLAuthorizationStatusDenied -> LocationGrantedType.NotAllowed
-            else -> LocationGrantedType.Denied
+            kCLAuthorizationStatusRestricted, kCLAuthorizationStatusDenied -> LocationPermissionResult.NotAllowed
+            else -> LocationPermissionResult.Denied
         }
     }
 
     override fun locationManager(manager: CLLocationManager, didFailWithError: NSError) {
         if (locationPermissionContinuation?.isActive == true) {
-            locationPermissionContinuation?.resume(LocationGrantedType.Denied)
+            locationPermissionContinuation?.resume(LocationPermissionResult.Denied)
             locationPermissionContinuation = null
         }